+++ src/index.ts
@@ -1,2 +1,1 @@
 export * from "./_internal/store";
-export * from "./_internal/types";