+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
-## 0.0.2 / 2023-06-xx
-- Upgrade to `mingo@6.4.0` for full update operator support.
+## 0.0.2 / 2023-06-08
+- Upgrade to `mingo@6.4.1` for full update operator support.
+- Default to not cloning inputs on update.
 
 ## 0.0.1 / 2023-06-04
 - Moved React integration to separate package [reack-adaka](https://www.npmjs.com/package/react-adaka).